Added getLayoutDefinition method to GeoTrellisUtils
Fixed merge conflicts in TileLayerMetadata
